- added Day 6 input generation code - added Day 6 generated input - fixed bugs in Day 6 code that were uncovered by the generated input - updated README with Day 6 expected values

+++ python/day_6_part_1.py
@@ -8,8 +8,7 @@ def solution(scenario: str):
 
     lines = np.array([line.strip().split() for line in input.strip().splitlines()])
     numerical_data = np.array([[int(d) for d in line] for line in lines
-                               if line[0] not in ["+", "*"]]).T
-
+                               if line[0] not in ["+", "*"]], dtype=np.int64).T
     result = 0
     for op, nums in zip(lines[-1], numerical_data):
         if op == "*":
